requirements/local.txt
-r base.txt
django-debug-toolbar==2.2
django-extensions==2.2.8
Faker==4.0.2